chore: move DirectoryTable inside DirectoryView
