+++ src/client/components/DirectoryView/components/DirectoryTable/components/DirectoryTableRow.tsx
@@ -1,9 +1,9 @@
 import { type MouseEvent, useContext } from "react";
 import { useRouter } from "next/router";
 import { flexRender, type Row } from "@tanstack/react-table";
-import { DirectoryTableContext } from "~/client/components/DirectoryTable/contexts/DirectoryContext";
-import { type DirectoryTableRowData } from "~/client/components/DirectoryTable/types";
-import { DirectoryContext } from "~/client/contexts/DirectoryContext";
+import { DirectoryTableContext } from "~/client/components/DirectoryView/components/DirectoryTable/contexts/DirectoryTableContext";
+import { type DirectoryTableRowData } from "~/client/components/DirectoryView/components/DirectoryTable/types";
+import { DirectoryContext } from "~/client/components/DirectoryView/contexts/DirectoryContext";
 import { clearWindowSelection } from "~/client/utils/clearWindowSelection";
 import { cx } from "~/client/utils/cx";
 import { mapObject } from "~/client/utils/mapObject";
