+++ src/client/components/DirectoryView/components/DirectoryTable/components/FileIcon.tsx
